Fix off by one error in anaglyph.go
We were steping one pixel too far on both the X and Y axis when calculating the anaglyph image.

This had no negative effects because image.At and image.Set silently ignore out of bounds access, but it was still wrong. This commit fixes the issue by ensuring we only step to the last pixel